backend/chart: add latest rates timestamps in chart data
If one of the coin/rates pairs is older than 2hs the backend set the
`chartIsUpToDate` flag to false inside the `account-summary` return
struct. This adds a map to the struct that list the latest update
timestamp for each coin. This will allow the frontend to show a
meaningful message in the future instead of just keep loading the chart.