frontend: fix various tests
some tests are failing due to our migration from CRA to vite.

Most tests that fail are due to our test trying to assert
an exact match of class name(s). Also some others are
are due to incorrect "usage" of async test, which seems to be ok
in CRA but not vite.