frontend: add/fix loading state + minor refactor
adding the QR loading  spinner "as is" without conditional
rendering, as it seems to unmount the video component when
a state changes in the component (scan-qr-video).

As there's no loading state, the spinner appears initially before
the QR reader starts with the video. Once the video loads,
the spinner disappears on its own.

Also created/put back `ScanQRDialog`.