frontend: fix slow camera detection
On macOS M1/M2 the whole send screen takes a long time to load
and QrScanner icon is even slower. This is an indication that
qr-scanner hasCamera takes very long to test if there are any
available cameras.

https://github.com/nimiq/qr-scanner/blob/abcfe1bce2703721408d8ce7ebde94a359998506/src/qr-scanner.ts#L13-L54

Unfortunatelly it looks like that enumerateDevices is for some
reason very slow.

     await navigator.mediaDevices.enumerateDevices();

The app used to check if there are any camera devices and only
shows the camera icon if some cameras have been detected.

This change removes the check and now always shows the Qr code
scanner icon.