frontend: send payment message to pocket widget
Added a new message that sends the transaction id to the widget
after the user signed. And a new message that sends info about
canceled payment.

With this Pocket can give immediate feedback in the widget.

Adding a new field correlationId to every message that allows to
correlate request and response messages.

Updated request-address@0.0.11 npm module.